Refactored Config to separate file. Refactored portal effect changing to separate file. Fixed server synced settings not working properly.
